Fixed a couple bugs in the bittorrent-sync.install file. The if statement at the beginning of post_install() function was lacking quotes around the grep call. This caused an error to occur when this call returned an empty string.
Also the comparison operator was set to "!=" when it should have been set to "==".
Now the useradd command will run when there is no user by the name "btsync" in the passwd file.
Rather than when there IS a user by such a name.

Also updated the sha256sum in the PKGBUILD file.